add-friends-likes first commit

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -8,18 +8,12 @@ import javax.validation.constraints.PastOrPresent;
 import javax.validation.constraints.Positive;
 import javax.validation.constraints.Size;
 import java.time.LocalDate;
+import java.util.Set;
 
-/**
- * целочисленный идентификатор;
- * название;
- * описание;
- * дата релиза;
- * продолжительность фильма.
- */
 @Data
 @AllArgsConstructor
 public class Film {
-    private int id;
+    private Long id;
 
     @NotBlank
     private String name;
