+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -7,6 +7,7 @@ import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Past;
 import java.time.LocalDate;
+import java.util.Set;
 
 /**
  * целочисленный идентификатор;
@@ -19,7 +20,7 @@ import java.time.LocalDate;
 @AllArgsConstructor
 public class User {
 
-    private int userId;
+    private Long userId;
 
     @Email
     private String email;
